Wrote app functionality
Wrote out code in the onButtonClick method that makes the app work when the user clicks on the Generate button. It calls the QR code API and grabs the link the user has entered into the TextField box, then saves the .png in a custom size temporarily. After it applies that image to the program itself, it deletes the temporary image from the directory to save space. Error catching is added as well for when an invalud input is entered, or if something breaks when trying to call the API.